Restructuring Gameta, moving out CLI and context code out of __init__
