+++ gameta/apply.py
@@ -3,14 +3,15 @@ from typing import List
 
 import click
 
-from . import gameta_cli, gameta_context, GametaContext
+from .cli import gameta_cli
+from .context import gameta_context, GametaContext
 
 
 __all__ = ['apply']
 
 
 @gameta_cli.command()
-@click.option('--command', '-c', type=str, required=True, help='Command to be executed')
+@click.option('--command', '-c', 'commands', type=str, required=True, multiple=True, help='Command to be executed')
 @click.option('--tags', '-t', type=str, multiple=True, default=[], help='Repository tags to apply commands to')
 @click.option('--repositories', '-r', type=str, multiple=True, default=[], help='Repositories to apply commands to')
 @click.option('--verbose', '-v', is_flag=True, default=False, help='Display execution output when command is applied')
@@ -20,7 +21,7 @@ __all__ = ['apply']
 @gameta_context
 def apply(
         context: GametaContext,
-        command: str,
+        commands: List[str],
         tags: List[str],
         repositories: List[str],
         verbose: bool,
@@ -32,7 +33,7 @@ def apply(
     \f
     Args:
         context (GametaContext): Gameta Context
-        command (str): CLI command to be applied
+        commands (List[str]): CLI command to be applied
         tags (List[str]): Repository tags to apply command to
         repositories (List[str]): Repositories to apply command to
         verbose (bool): Flag to indicate that output should be displayed as the command is applied
@@ -56,11 +57,11 @@ def apply(
             )
         )
         click.echo(
-            f"Applying '{command}' to repos {repos if repos else list(context.repositories.keys())}"
+            f"Applying '{commands}' to repos {repos if repos else list(context.repositories.keys())}"
             f"{' in a separate shell' if shell else ''}"
         )
-        for repo, c in context.apply([command], repos=repos, shell=shell):
-            click.echo(f"Executing {command} in {repo}")
+        for repo, c in context.apply(commands, repos=repos, shell=shell):
+            click.echo(f"Executing {' '.join(c)} in {repo}")
             try:
                 if verbose:
                     click.echo(subprocess.check_output(c))
@@ -69,10 +70,10 @@ def apply(
             except subprocess.SubprocessError as e:
                 if raise_errors:
                     raise click.ClickException(
-                        f'Error {e.__class__.__name__}.{str(e)} occurred when executing command {command} in {repo}'
+                        f'Error {e.__class__.__name__}.{str(e)} occurred when executing command {" ".join(c)} in {repo}'
                     )
                 click.echo(
-                    f'Error {e.__class__.__name__}.{str(e)} occurred when executing command {command} in {repo}, '
+                    f'Error {e.__class__.__name__}.{str(e)} occurred when executing command {" ".join(c)} in {repo}, '
                     f'continuing execution'
                 )
     except click.ClickException:
