+++ pyproject.toml
@@ -35,4 +35,4 @@ tox = "^3.16.1"
 tox-pyenv = "^1.1.0"
 
 [tool.poetry.scripts]
-gameta = "gameta:gameta_cli"
+gameta = "gameta.cli:gameta_cli"
