+++ tests/test_apply.py
@@ -1,3 +1,4 @@
+import json
 import zipfile
 from os.path import join, dirname
 from shutil import copyfile, copytree
@@ -5,8 +6,8 @@ from unittest import TestCase
 from unittest.mock import patch
 
 from click.testing import CliRunner
-from gameta import GametaContext
 
+from gameta.context import GametaContext, SHELL
 from gameta.apply import apply
 
 
@@ -15,7 +16,7 @@ class TestApply(TestCase):
         self.runner = CliRunner()
         self.apply = apply
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_apply_key_parameters_not_provided(self, mock_ensure_object):
         with self.runner.isolated_filesystem() as f:
             with zipfile.ZipFile(join(dirname(__file__), 'data', 'git.zip'), 'r') as template:
@@ -35,10 +36,10 @@ class TestApply(TestCase):
                 "Error: Missing option '--command' / '-c'.\n"
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_apply_command_to_all_repositories(self, mock_ensure_object):
         params = {
-            'command': 'git fetch --all --tags --prune'
+            'commands': ('git fetch --all --tags --prune', ),
         }
         with self.runner.isolated_filesystem() as f:
             copytree(join(dirname(dirname(__file__)), '.git'), join(f, '.git'))
@@ -51,11 +52,11 @@ class TestApply(TestCase):
             context.project_dir = f
             context.load()
             mock_ensure_object.return_value = context
-            result = self.runner.invoke(self.apply, ['--command', params['command']])
+            result = self.runner.invoke(self.apply, ['--command', params['commands'][0]])
             self.assertEqual(result.exit_code, 0)
             self.assertEqual(
                 result.output,
-                f"Applying '{params['command']}' to repos ['gameta', 'GitPython', 'gitdb']\n"
+                f"Applying '{params['commands']}' to repos ['gameta', 'GitPython', 'gitdb']\n"
                 "Executing git fetch --all --tags --prune in gameta\n"
                 "Executing git fetch --all --tags --prune in GitPython\n"
                 "Executing git fetch --all --tags --prune in gitdb\n"
@@ -64,7 +65,7 @@ class TestApply(TestCase):
     @patch('gameta.click.Context.ensure_object')
     def test_apply_command_to_tagged_repositories(self, mock_ensure_object):
         params = {
-            'command': 'git fetch --all --tags --prune',
+            'commands': ('git fetch --all --tags --prune', ),
             'tags': ('c', )
         }
         with self.runner.isolated_filesystem() as f:
@@ -78,19 +79,19 @@ class TestApply(TestCase):
             context.project_dir = f
             context.load()
             mock_ensure_object.return_value = context
-            result = self.runner.invoke(self.apply, ['--command', params['command'], '-t', params['tags'][0]])
+            result = self.runner.invoke(self.apply, ['--command', params['commands'][0], '-t', params['tags'][0]])
             self.assertEqual(result.exit_code, 0)
             self.assertEqual(
                 result.output,
-                f"Applying '{params['command']}' to repos ['GitPython', 'gitdb']\n"
+                f"Applying '{params['commands']}' to repos ['GitPython', 'gitdb']\n"
                 "Executing git fetch --all --tags --prune in GitPython\n"
                 "Executing git fetch --all --tags --prune in gitdb\n"
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_apply_command_to_tagged_repositories_that_do_not_exist(self, mock_ensure_object):
         params = {
-            'command': 'git fetch --all --tags --prune',
+            'commands': ('git fetch --all --tags --prune', ),
             'tags': ('hello', 'world')
         }
         with self.runner.isolated_filesystem() as f:
@@ -105,21 +106,21 @@ class TestApply(TestCase):
             context.load()
             mock_ensure_object.return_value = context
             result = self.runner.invoke(
-                self.apply, ['--command', params['command'], '-t', params['tags'][0], '-t', params['tags'][1]]
+                self.apply, ['--command', params['commands'][0], '-t', params['tags'][0], '-t', params['tags'][1]]
             )
             self.assertEqual(result.exit_code, 0)
             self.assertEqual(
                 result.output,
-                f"Applying '{params['command']}' to repos ['gameta', 'GitPython', 'gitdb']\n"
+                f"Applying '{params['commands']}' to repos ['gameta', 'GitPython', 'gitdb']\n"
                 "Executing git fetch --all --tags --prune in gameta\n"
                 "Executing git fetch --all --tags --prune in GitPython\n"
                 "Executing git fetch --all --tags --prune in gitdb\n"
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_apply_command_to_specified_repositories(self, mock_ensure_object):
         params = {
-            'command': 'git fetch --all --tags --prune',
+            'commands': ('git fetch --all --tags --prune', ),
             'repositories': ('GitPython', )
         }
         with self.runner.isolated_filesystem() as f:
@@ -133,18 +134,20 @@ class TestApply(TestCase):
             context.project_dir = f
             context.load()
             mock_ensure_object.return_value = context
-            result = self.runner.invoke(self.apply, ['--command', params['command'], '-r', params['repositories'][0]])
+            result = self.runner.invoke(
+                self.apply, ['--command', params['commands'][0], '-r', params['repositories'][0]]
+            )
             self.assertEqual(result.exit_code, 0)
             self.assertEqual(
                 result.output,
-                f"Applying '{params['command']}' to repos ['GitPython']\n"
+                f"Applying '{params['commands']}' to repos ['GitPython']\n"
                 "Executing git fetch --all --tags --prune in GitPython\n"
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_apply_command_to_specified_repositories_that_do_not_exist(self, mock_ensure_object):
         params = {
-            'command': 'git fetch --all --tags --prune',
+            'commands': ('git fetch --all --tags --prune', ),
             'repositories': ('hello', 'world')
         }
         with self.runner.isolated_filesystem() as f:
@@ -160,21 +163,21 @@ class TestApply(TestCase):
             mock_ensure_object.return_value = context
             result = self.runner.invoke(
                 self.apply,
-                ['--command', params['command'], '-r', params['repositories'][0], '-r', params['repositories'][1]]
+                ['--command', params['commands'][0], '-r', params['repositories'][0], '-r', params['repositories'][1]]
             )
             self.assertEqual(result.exit_code, 0)
             self.assertEqual(
                 result.output,
-                f"Applying '{params['command']}' to repos ['gameta', 'GitPython', 'gitdb']\n"
+                f"Applying '{params['commands']}' to repos ['gameta', 'GitPython', 'gitdb']\n"
                 "Executing git fetch --all --tags --prune in gameta\n"
                 "Executing git fetch --all --tags --prune in GitPython\n"
                 "Executing git fetch --all --tags --prune in gitdb\n"
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_apply_command_to_merged_tags_and_repos(self, mock_ensure_object):
         params = {
-            'command': 'git fetch --all --tags --prune',
+            'commands': ('git fetch --all --tags --prune', ),
             'tags': ('metarepo', ),
             'repositories': ('gitdb', )
         }
@@ -190,20 +193,21 @@ class TestApply(TestCase):
             context.load()
             mock_ensure_object.return_value = context
             result = self.runner.invoke(
-                self.apply, ['--command', params['command'], '-t', params['tags'][0], '-r', params['repositories'][0]]
+                self.apply,
+                ['--command', params['commands'][0], '-t', params['tags'][0], '-r', params['repositories'][0]]
             )
             self.assertEqual(result.exit_code, 0)
             self.assertEqual(
                 result.output,
-                f"Applying '{params['command']}' to repos ['gameta', 'gitdb']\n"
+                f"Applying '{params['commands']}' to repos ['gameta', 'gitdb']\n"
                 "Executing git fetch --all --tags --prune in gameta\n"
                 "Executing git fetch --all --tags --prune in gitdb\n"
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_apply_command_in_separate_shell(self, mock_ensure_object):
         params = {
-            'command': 'git fetch --all --tags --prune',
+            'commands': ('git fetch --all --tags --prune', ),
         }
         with self.runner.isolated_filesystem() as f:
             copytree(join(dirname(dirname(__file__)), '.git'), join(f, '.git'))
@@ -216,20 +220,20 @@ class TestApply(TestCase):
             context.project_dir = f
             context.load()
             mock_ensure_object.return_value = context
-            result = self.runner.invoke(self.apply, ['--command', params['command'], '-s'])
+            result = self.runner.invoke(self.apply, ['--command', params['commands'][0], '-s'])
             self.assertEqual(result.exit_code, 0)
             self.assertEqual(
                 result.output,
-                f"Applying '{params['command']}' to repos ['gameta', 'GitPython', 'gitdb'] in a separate shell\n"
-                "Executing git fetch --all --tags --prune in gameta\n"
-                "Executing git fetch --all --tags --prune in GitPython\n"
-                "Executing git fetch --all --tags --prune in gitdb\n"
+                f"Applying '{params['commands']}' to repos ['gameta', 'GitPython', 'gitdb'] in a separate shell\n"
+                f"Executing {SHELL} -c  git fetch --all --tags --prune in gameta\n"
+                f"Executing {SHELL} -c  git fetch --all --tags --prune in GitPython\n"
+                f"Executing {SHELL} -c  git fetch --all --tags --prune in gitdb\n"
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_apply_raise_errors(self, mock_ensure_object):
         params = {
-            'command': 'git fetch --all --tags --prune'
+            'commands': ('git fetch --all --tags --prune', )
         }
         with self.runner.isolated_filesystem() as f:
             with zipfile.ZipFile(join(dirname(__file__), 'data', 'git.zip'), 'r') as template:
@@ -243,13 +247,13 @@ class TestApply(TestCase):
             context.project_dir = f
             context.load()
             mock_ensure_object.return_value = context
-            result = self.runner.invoke(self.apply, ['--command', params['command'], '-e'])
+            result = self.runner.invoke(self.apply, ['--command', params['commands'][0], '-e'])
             self.assertEqual(result.exit_code, 1)
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_apply_verbose(self, mock_ensure_object):
         params = {
-            'command': 'git fetch --all --tags --prune'
+            'commands': ('git fetch --all --tags --prune', )
         }
         with self.runner.isolated_filesystem() as f:
             copytree(join(dirname(dirname(__file__)), '.git'), join(f, '.git'))
@@ -262,11 +266,11 @@ class TestApply(TestCase):
             context.project_dir = f
             context.load()
             mock_ensure_object.return_value = context
-            result = self.runner.invoke(self.apply, ['--command', params['command'], '-v'])
+            result = self.runner.invoke(self.apply, ['--command', params['commands'][0], '-v'])
             self.assertEqual(result.exit_code, 0)
             self.assertEqual(
                 result.output,
-                f"Applying '{params['command']}' to repos ['gameta', 'GitPython', 'gitdb']\n"
+                f"Applying '{params['commands']}' to repos ['gameta', 'GitPython', 'gitdb']\n"
                 "Executing git fetch --all --tags --prune in gameta\n"
                 "Fetching origin\n"
                 "\n"
