+++ tests/test_init.py
@@ -9,7 +9,7 @@ from unittest.mock import patch
 
 from click.testing import CliRunner
 
-from gameta import GametaContext
+from gameta.context import GametaContext
 from gameta.init import init, sync
 
 
@@ -18,7 +18,7 @@ class TestInit(TestCase):
         self.runner = CliRunner()
         self.init = init
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_init_with_default_values_folder_is_not_a_git_repo(self, mock_ensure_object):
         with self.runner.isolated_filesystem() as f:
             context = GametaContext()
@@ -32,7 +32,7 @@ class TestInit(TestCase):
                 f"Error: {f} is not a valid git repo, initialise it with -g flag\n"
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_init_with_git_is_true_folder_is_not_a_git_repo(self, mock_ensure_object):
         with self.runner.isolated_filesystem() as f:
             context = GametaContext()
@@ -60,7 +60,7 @@ class TestInit(TestCase):
                     }
                 )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_init_with_default_values_folder_is_a_git_repo(self, mock_ensure_object):
         with self.runner.isolated_filesystem() as f:
             with zipfile.ZipFile(join(dirname(__file__), 'data', 'git.zip'), 'r') as template:
@@ -89,7 +89,7 @@ class TestInit(TestCase):
                     }
                 )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_init_with_default_values_folder_is_a_metarepo(self, mock_ensure_object):
         with self.runner.isolated_filesystem() as f:
             with zipfile.ZipFile(join(dirname(__file__), 'data', 'git.zip'), 'r') as template:
@@ -120,7 +120,7 @@ class TestInit(TestCase):
                     }
                 )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_init_with_overwrite_folder_is_a_metarepo(self, mock_ensure_object):
         with self.runner.isolated_filesystem() as f:
             with zipfile.ZipFile(join(dirname(__file__), 'data', 'git.zip'), 'r') as template:
@@ -157,7 +157,7 @@ class TestSync(TestCase):
         self.runner = CliRunner()
         self.sync = sync
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_sync_empty_meta_file(self, mock_ensure_object):
         with self.runner.isolated_filesystem() as f:
             with zipfile.ZipFile(join(dirname(__file__), 'data', 'git.zip'), 'r') as template:
@@ -179,7 +179,7 @@ class TestSync(TestCase):
                 f'Syncing all child repositories in metarepo {f}\n'
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_sync_all_repos(self, mock_ensure_object):
         with self.runner.isolated_filesystem() as f:
             with zipfile.ZipFile(join(dirname(__file__), 'data', 'git.zip'), 'r') as template:
