+++ tests/test_parameters.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 
 from click.testing import CliRunner
 
-from gameta import GametaContext
+from gameta.context import GametaContext
 from gameta.parameters import add, delete
 
 
@@ -15,7 +15,7 @@ class TestAdd(TestCase):
         self.runner = CliRunner()
         self.add = add
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_add_parameters_missing_key_parameters(self, mock_ensure_object):
         with self.runner.isolated_filesystem() as f:
             copyfile(join(dirname(__file__), 'data', '.meta_other_repos'), join(f, '.meta'))
@@ -33,7 +33,7 @@ class TestAdd(TestCase):
                 "Error: Missing option '--parameter' / '-p'.\n"
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_add_parameters_skip_user_prompt_default_values(self, mock_ensure_object):
         params = {
             'parameter': 'test'
@@ -81,7 +81,7 @@ class TestAdd(TestCase):
                     }
                 )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_add_parameters_skip_user_prompt_user_provided_default_value(self, mock_ensure_object):
         params = {
             'parameter': 'test',
@@ -130,7 +130,7 @@ class TestAdd(TestCase):
                     }
                 )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_add_parameters_skip_user_prompt_default_type_not_in_choice(self, mock_ensure_object):
         params = {
             'parameter': 'test',
@@ -153,7 +153,7 @@ class TestAdd(TestCase):
                 "(choose from int, float, str, bool, dict, list)\n"
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_add_parameters_user_prompt_all_value_prompted(self, mock_ensure_object):
         params = {
             'parameter': 'test',
@@ -211,7 +211,7 @@ class TestAdd(TestCase):
                     }
                 )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_add_parameters_user_prompt_skipping_with_user_provided_default_value(self, mock_ensure_object):
         params = {
             'parameter': 'test',
@@ -270,7 +270,7 @@ class TestAdd(TestCase):
                     }
                 )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_add_parameters_user_prompt_complex_user_input(self, mock_ensure_object):
         params = {
             'parameter': 'test',
@@ -330,7 +330,7 @@ class TestAdd(TestCase):
                     }
                 )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_add_parameters_user_prompt_user_input_does_not_match_required_type(self, mock_ensure_object):
         params = {
             'parameter': 'test',
@@ -398,7 +398,7 @@ class TestDelete(TestCase):
         self.runner = CliRunner()
         self.delete = delete
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_add_parameters_missing_key_parameters(self, mock_ensure_object):
         with self.runner.isolated_filesystem() as f:
             copyfile(join(dirname(__file__), 'data', '.meta_other_repos'), join(f, '.meta'))
@@ -416,7 +416,7 @@ class TestDelete(TestCase):
                 "Error: Missing option '--parameter' / '-p'.\n"
             )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_delete_parameters_parameter_does_not_exist(self, mock_ensure_object):
         params = {
             'parameter': 'test'
@@ -458,7 +458,7 @@ class TestDelete(TestCase):
                     }
                 )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_delete_parameters_all_parameters_deleted(self, mock_ensure_object):
         params = {
             'parameter': 'test'
@@ -506,7 +506,7 @@ class TestDelete(TestCase):
                     }
                 )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_delete_parameters_partial_parameter_deleted(self, mock_ensure_object):
         params = {
             'parameter': 'test'
@@ -553,7 +553,7 @@ class TestDelete(TestCase):
                     }
                 )
 
-    @patch('gameta.click.Context.ensure_object')
+    @patch('gameta.cli.click.Context.ensure_object')
     def test_delete_parameters_attempting_to_delete_reserved_parameters(self, mock_ensure_object):
         params = {
             'parameter': 'url'
